Refactor dstack ls
