+++ cli/dstack/utils/common.py
@@ -85,3 +85,11 @@ def _relative_timestamp_to_datetime(amount, unit):
         "w": 7 * 24 * 3600,
     }[unit]
     return datetime.utcnow() + timedelta(seconds=amount * multiplier * -1)
+
+
+def sizeof_fmt(num, suffix="B"):
+    for unit in ["", "Ki", "Mi", "Gi", "Ti", "Pi", "Ei", "Zi"]:
+        if abs(num) < 1024.0:
+            return "%3.1f%s%s" % (num, unit, suffix)
+        num /= 1024.0
+    return "%.1f%s%s" % (num, "Yi", suffix)
